+++ src/main/java/mod/gottsch/forge/treasure2/core/entity/TreasureEntities.java
@@ -18,11 +18,7 @@
 package mod.gottsch.forge.treasure2.core.entity;
 
 
-import mod.gottsch.forge.treasure2.core.entity.monster.BoundSoul;
-import mod.gottsch.forge.treasure2.core.entity.monster.CauldronChestMimic;
-import mod.gottsch.forge.treasure2.core.entity.monster.PirateChestMimic;
-import mod.gottsch.forge.treasure2.core.entity.monster.VikingChestMimic;
-import mod.gottsch.forge.treasure2.core.entity.monster.WoodChestMimic;
+import mod.gottsch.forge.treasure2.core.entity.monster.*;
 import mod.gottsch.forge.treasure2.core.setup.Registration;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.MobCategory;
@@ -38,6 +34,8 @@ public class TreasureEntities {
 	public static final String PIRATE_CHEST_MIMIC = "pirate_chest_mimic";
 	public static final String VIKING_CHEST_MIMIC = "viking_chest_mimic";
 	public static final String CAULDRON_CHEST_MIMIC = "cauldron_chest_mimic";
+	public static final String CRATE_CHEST_MIMIC = "crate_chest_mimic";
+	public static final String MOLDY_CRATE_CHEST_MIMIC = "moldy_crate_chest_mimic";
 	
 	public static final RegistryObject<EntityType<BoundSoul>> BOUND_SOUL_ENTITY_TYPE = Registration.ENTITIES.register(BOUND_SOUL, () -> EntityType.Builder.of(BoundSoul::new, MobCategory.MONSTER)
 			.sized(0.6F, 1.95F)
@@ -73,6 +71,20 @@ public class TreasureEntities {
 			.setTrackingRange(80)
 			.setShouldReceiveVelocityUpdates(true)
 			.build(CAULDRON_CHEST_MIMIC));
+	
+	public static final RegistryObject<EntityType<CrateChestMimic>> CRATE_CHEST_MIMIC_ENTITY_TYPE = Registration.ENTITIES.register(CRATE_CHEST_MIMIC, () -> EntityType.Builder.of(CrateChestMimic::new, MobCategory.MONSTER)
+			.sized(0.875F, 0.875F)
+			.clientTrackingRange(12)
+			.setTrackingRange(80)
+			.setShouldReceiveVelocityUpdates(true)
+			.build(CRATE_CHEST_MIMIC));
+	
+	public static final RegistryObject<EntityType<MoldyCrateChestMimic>> MOLDY_CRATE_CHEST_MIMIC_ENTITY_TYPE = Registration.ENTITIES.register(MOLDY_CRATE_CHEST_MIMIC, () -> EntityType.Builder.of(MoldyCrateChestMimic::new, MobCategory.MONSTER)
+			.sized(0.875F, 0.875F)
+			.clientTrackingRange(12)
+			.setTrackingRange(80)
+			.setShouldReceiveVelocityUpdates(true)
+			.build(MOLDY_CRATE_CHEST_MIMIC));
 	
 	public static void register() {
 		// cycle through all block and create items